Addressing discrepancies between the dist version of WP and wordpress-develop (#30)
Some files have been moved around in the `trunk` version of wordpress-develop, which has caused the test suite to fail. This means we need to download the corresponding version of wordpress-develop as the "dist" version of WordPress we're testing against.

Further issues are caused by wordpress-develop using an ever so slightly different numbering scheme from the "dist" version, as wordpress-develop would call version 7.6 7.6.0, adding a zero as the patch component.

Note that this removes support for testing against the `trunk`/`main`/`master` for the time being.